Mitigate deadlock on remote write
Fix and mitigate possible issues when continuously writing to slow a
receiver. The syntomps are a deadlocked qemu machine which is waiting
for a mutex own by the JSON remote memory writer.
The cause is a over-pressured QemuSocket which is overflowing the
sending buffer and waiting for confirmation of unsent/lost fragments.

Signed-off-by: Luca Bruno <lucab@debian.org>